+++ packages/language-service/BUILD.bazel
@@ -19,7 +19,6 @@ ts_library(
         "//packages/compiler",
         "//packages/compiler-cli",
         "//packages/core",
-        "//packages/language-service/common",
         "@npm//@types/node",
         "@npm//typescript",
     ],
