--- packages/language-service/common/BUILD.bazel
@@ -1,11 +1,0 @@
-load("//tools:defaults.bzl", "ts_library")
-
-package(default_visibility = ["//packages/language-service:__subpackages__"])
-
-ts_library(
-    name = "common",
-    srcs = glob(["*.ts"]),
-    deps = [
-        "@npm//typescript",
-    ],
-)
