+++ packages/language-service/ivy/BUILD.bazel
@@ -17,7 +17,6 @@ ts_library(
         "//packages/compiler-cli/src/ngtsc/shims",
         "//packages/compiler-cli/src/ngtsc/typecheck",
         "//packages/compiler-cli/src/ngtsc/typecheck/api",
-        "//packages/language-service/common",
         "@npm//typescript",
     ],
 )
