+++ packages/language-service/test/utils_spec.ts
@@ -9,8 +9,7 @@
 import * as ng from '@angular/compiler';
 import * as ts from 'typescript';
 
-import {getClassDeclFromDecoratorProp} from '../common/ts_utils';
-import {getDirectiveClassLike} from '../src/ts_utils';
+import {getClassDeclFromDecoratorProp, getDirectiveClassLike} from '../src/ts_utils';
 import {getPathToNodeAtPosition} from '../src/utils';
 import {MockTypescriptHost} from './test_utils';
